Move static functions and their declarations to a separate file
